+++ CHANGELOG.md
@@ -37,6 +37,7 @@
   - fixed automatic upscrolling in form pages
 - Service Release Process
   - fixed automatic upscrolling in form pages
+  - Added conformity document UI and functionality
 
 ## 1.7.0